Move sourcelink to utils folder
